Fixed: stop guessing from file extension
I don't exactly remember why I did that but it definitely shouldn't be
the generic way to do things. It can lead to unexpected behavior and
increase risks of crashes in __init__() of decompressors. Crashing at
this place is not resumable.

Therefore, if you really want this feature, simply implement it in your
own decompressor class.